Move salesforcedx-vscode-apex tests to vscode-integration dir
